+++ packet-atlas/patches/backups/App.before-v3.2-repair.20260604150858.tsx
@@ -1,7 +1,0 @@
-import { PacketAtlasPage } from './features/packet-atlas/PacketAtlasPage'
-
-function App() {
-  return <PacketAtlasPage />
-}
-
-export default App